# clingo-bootsrap: pgo, lto, mimalloc optimizations