CMake has poor support for PGO, so I put it in our build script. Which
only works for %gcc, since clang needs a merge step with `llvm-profdata`
after generating reports.


```
$ spack solve --timers hdf5
    setup          4.231s
    load           0.016s
    ground         2.039s
    solve          3.155s
    total          9.464s

$ spack install clingo-bootstrap +optimized  # installs clingo-bootstrap/6wg272k

$ spack load /6wg272k

$ spack solve --timers hdf5
    setup          4.179s
    load           0.014s
    ground         1.435s
    solve          2.279s
    total          7.930s
```