@haampie 2023-01-15
Some more timings:

**system python 3.10.7, clingo-bootstrap@spack official binaries**

    setup          4.113s
    load           0.018s
    ground         1.848s
    solve          2.523s
    total          8.698s

**clingo-bootstrap@spack ~optimizations %gcc ^python@3.11 +optimizations**

    setup          4.163s
    load           0.018s
    ground         1.812s
    solve          2.518s
    total          8.707s

**clingo-bootstrap@spack +optimizations %clang ^python@3.11 +optimizations**

    setup          3.389s
    load           0.018s
    ground         1.247s
    solve          1.792s
    total          6.615s

**clingo-bootstrap@spack +optimizations %gcc ^python@3.11 +optimizations**

    setup          3.372s
    load           0.015s
    ground         1.234s
    solve          1.872s
    total          6.657s

**clingo-bootstrap@5.6.2 +optimizations %clang ^python@3.11 +optimizations**

    setup          3.965s
    load           0.015s
    ground         1.434s
    solve          1.798s
    total          7.393s

**clingo-bootstrap@5.6.2 +optimizations %gcc ^python@3.11 +optimizations**

    setup          3.951s
    load           0.015s
    ground         1.567s
    solve          1.870s
    total          7.582s