@haampie 2023-01-16
@trws fyi, mimalloc works "by chance" if you dynamically link libstdc++.so: operator new/delete symbols have to be located in libclingo first, so if we would use another Python extension that links to libstdc++ and it gets imported earlier than clingo, the operator new/delete symbols are no longer overridden.

For bootstrapping though, we statically link libstdc++, so all is good. (Well, not 100% if we control the order on the link line `-lmimalloc -lstdc++`, but right now that is the case, fortunately)