@haampie 2023-02-18
I've updated this locally with changes for macOS, but without mimalloc.