@haampie 2023-08-09
Might be worth bisecting Spack's commits. The 20 - 30% reduction in solve time still exists today, but hdf5 solve times have double between 68ce49a57b5d8727fa0503d97cf3def3462c44e1 then and f9457fa80ba51d14610df8b57a6b5a514d6bb6ff now.