@haampie 2023-08-10
More speed comparisons:

```
 Row │ pkg            before   after    speedup 
     │ String15       Float64  Float64  Float64 
─────┼──────────────────────────────────────────
   1 │ aluminum        17.425   13.748    1.267
   2 │ ascent          37.138   30.863    1.203
   3 │ axom            35.695   29.129    1.225
   4 │ blt             14.406   11.585    1.244
   5 │ caliper         18.247   15.092    1.209
   6 │ care            15.591   12.941    1.205
   7 │ chai            15.409   12.514    1.231
   8 │ conduit         20.319   17.541    1.158
   9 │ dihydrogen      19.39    15.956    1.215
  10 │ flux-core       16.974   13.848    1.226
  11 │ flux-sched      17.703   14.135    1.252
  12 │ flux-security   15.273   12.622    1.21
  13 │ glvis           35.697   28.373    1.258
  14 │ hydrogen        17.196   14.357    1.198
  15 │ hypre           23.511   19.515    1.205
  16 │ lbann           34.321   27.585    1.244
  17 │ lvarray         19.693   16.001    1.231
  18 │ mfem            33.937   28.696    1.183
  19 │ py-hatchet      25.908   20.835    1.243
  20 │ py-maestrowf    18.127   14.746    1.229
  21 │ py-merlin       19.904   15.984    1.245
  22 │ py-shroud       15.383   12.473    1.233
  23 │ raja            15.219   12.199    1.248
  24 │ samrai          19.468   15.218    1.279
  25 │ scr             19.318   14.962    1.291
  26 │ sundials        30.22    25.351    1.192
  27 │ umpire          14.815   11.796    1.256
  28 │ visit           34.759   28.839    1.205
  29 │ xbraid          18.273   14.179    1.289
  30 │ zfp             13.953   11.151    1.251
```

seems pretty consistent.