@haampie 2023-08-10
Regarding mimalloc: I think our best bet is to (a) statically link `libstdc++.so`, (b) use the mimalloc object file, and (c) ensure all symbols are local except for `clingo_*` etc.

With shared libstdc++ other override tricks may not work because they depend on the load order.

Alternatively, using a custom allocator in clingo is a major refactor, not great...