@trws 2023-08-10
FWIW, truly static libc++ is much easier than truly static libstdc++. Not sure if we expose it, but there's literally a flag on the llvm cmake to build a truly static libc++, makes it a lot nicer to deal with.