@haampie 2023-08-10
Also FWIW, the issue being solved is mainly how to override `operator new/delete` _locally_. Some approaches:

1. `libclingo.so` with mimalloc `operator new` and shared libstdc++ works generally (libstdc++ resolves internal operator new also in libclingo.so), but (a) if it works it overrides `operator new` globally, (b) it depends on load order, which ultimately comes down to python import order.
2. `libclingo.so` with mimalloc `operator new` and static libstdc++ also seems to work, except that mimalloc static lib contains symbols with global visibility (maybe I'm missing a linker flag to avoid that), so other libraries may end up using mimalloc too
3. `libclingo.so` with mimalloc `operator new` and static libstdc++ and a version/linker script that hides all non-clingo symbols; should be robust