@trws 2023-08-10
Option 3, 100% no question, whichever C++ stdlib it is.  That's the approach we use for a lot of tool libs, and the approach used by nvidia for libcudart.  It's reliable, and keeps everything segregated.  The only trick will be if we find out that something does unclean things with the system allocator between libclingo and outside.  Here's hoping not.

It's worth noting that there are also multiple memory pools in libstdc++, the string one has bitten me in the past when trying to do things like this.  Not sure if libc++ has the same problem or not.