@haampie 2023-08-10
Nice, so, yeah, that's what the PR currently does :upside_down_face: 

I could also add your upstream hash function improvements as a patch for the latest clingo version.