@haampie 2023-08-12
I've upstreamed the version script change and [that got merged](https://github.com/potassco/clingo/pull/447)