@rkaminsk 2023-08-13
@haampie not sure exactly how spack works but if you are processing a lot of symbols, there might also be some overhead simply because a lot of functions a called. Standard python does not inline functions. I have sometimes seen huge performance improvements switching to pypy. At the expense of ugly code, one could probably also do the inlining manually.