@haampie 2023-08-13
I believe that's the reason we're stuck at commit https://github.com/potassco/clingo/commit/2a025667090d71b2c9dce60fe924feb6bde8f667, one commit before cffi. With cffi it's about a 10% performance regression due to many C calls, if I remember correctly.