@rkaminsk 2023-08-13
> I believe that's the reason we're stuck at commit [potassco/clingo@2a02566](https://github.com/potassco/clingo/commit/2a025667090d71b2c9dce60fe924feb6bde8f667), one commit before cffi. With cffi it's about a 10% performance regression due to many C calls, if I remember correctly.

There is still a bit of optimization potential here by inlining `_c_call` and `_c_call2` from the `_internal` module in critical places. Also checking the result before calling `_handle_error` should give a small boost. There is a penalty as compared to the previous implementation. Classes implemented via the C-API have less overhead as compared to classes declared in Python. However, the maintenance burden of an extension based on CPython is just too high. Maybe Cython would be a faster alternative. At least for the standard Python implementation...